Close #2 for now; css grid & margin auto center nicely
It got overly complex with the previous method of
	transform: translate(-50%, -50%)
because it created a scrolling issue doesn't need to exist.
This was also a non-problem for mobile prior to adding all this
CSS, and just made everything lame.  Simpler solution, which
works fine for now is...
	body { display: grid; }
	canvas { margin: auto; }
which effectively centers it within the body element, and is
good enough and is simple.  Reopen later only if you are feeling
REALLY particular about how the zoom/scroll should work.